Improved error reporting in case of a SD card read error. Return SPI statusbyte too.

+++ firmware/sdcard/Sd2Card.h
@@ -183,7 +183,7 @@ class Sd2Card {
   // int errorCode() const {return errorCode_;}
 
   /** \return error data for last error. */
-  int errorData() const {return status_;}
+  uint8_t errorData() const {return status_;}
   /**
    * Initialize an SD flash memory card with default clock rate and chip
    * select pin.  See sd2Card::init(uint8_t sckRateID, uint8_t chipSelectPin).
